requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.9.1
altgraph==0.17.3
aniso8601==7.0.0
arrow==1.2.3
# asgiref==3.5.0
async-timeout==4.0.2
attrs==22.1.0
auto-py-to-exe==2.26.0
autopep8==2.0.0
Babel==2.12.1
bcrypt==3.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
blinker==1.5
boto3==1.28.35
bottle==0.12.23
bottle-websocket==0.2.9
bs4==0.0.1
cachetools==5.2.1
certifi==2022.9.24
cffi==1.15.0
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.2
cloud-sql-python-connector==1.1.0
colorama==0.4.6
contourpy==1.0.6
cookiecutter==2.1.1
coverage==7.0.3
cryptography==39.0.0
cycler==0.11.0
dearpygui==1.8.0
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
dnspython==2.2.1
docopt==0.6.2
dominate==2.7.0
Eel==0.14.0
email-validator==1.1.3
et-xmlfile==1.1.0
exceptiongroup==1.0.4
filelock==3.8.0
Flask==2.2.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-JWT
Flask-JWT-Extended
Flask-Login==0.6.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.1
Flask-Minify==0.41
Flask-MySQLdb==1.0.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
fonttools==4.38.0
frozenlist==1.3.3
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
google-auth==2.15.0
google-auth-oauthlib==0.8.0
graphene==2.1.9
graphene-django==2.14.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet
gspread==5.7.2
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.1.1
jinja2-time==0.2.0
joblib==1.2.0
jsmin==3.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
lesscpy==0.15.1
logzero==1.7.0
lxml==4.9.1
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
ndg-httpsclient==0.5.1
nltk==3.7
numpy==1.23.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==22.0
pandas==1.5.0
passlib==1.7.4
pefile==2022.5.30
phonenumbers==8.13.7
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.3
pluggy==1.0.0
ply==3.11
promise==2.3
protobuf==3.20.1
psycogreen==1.0.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pusher==3.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pygame==2.3.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyJWT
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
# pypiwin32==223
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.3
pytest==7.2.0
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest-timeit==0.3.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==7.0.0
python-stripe==0.0.6
pytrends==4.8.0
pytz==2022.2.1
pywin32-ctypes==0.2.0
PyYAML==6.0
qiniu==7.10.0
qtgui==0.0.1
rcssmin==1.1.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
soupsieve==2.3.2.post1
speaklater==1.3
SQLAlchemy==1.4.35
sqlparse==0.4.2
stripe==5.2.0
text-unidecode==1.3
textblob==0.17.1
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.6
upload==0.1.1
urllib3==1.26.12
virtualenv==20.16.6
visitor==0.1.3
Werkzeug==2.2.2
whichcraft==0.6.1
WTForms==3.0.1
wtforms-html5==0.6.1
xlrd==2.0.1
xlwings==0.28.7
xxhash==3.2.0
yarg==0.1.9
yarl==1.8.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.5.2
